requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
gspread
oauth2client
fpdf
langchain-groq
langchain-core
groq
reportlab